requirements.txt
fastapi==0.114.0

uvicorn==0.30.6

PyJWT==2.9.0

sqlalchemy==2.0.35

SQLAlchemy==2.0.35
